+++ Laser_Tag/src/mc/cyberplex/LaserTag/arena/Message.java
@@ -1,0 +1,5 @@
+package mc.cyberplex.LaserTag.arena;
+
+public enum Message {
+	LOBBY, GAME
+}
