+++ Laser_Tag/src/mc/cyberplex/LaserTag/kits/enumerations/Colors.java
@@ -1,0 +1,5 @@
+package mc.cyberplex.LaserTag.kits.enumerations;
+
+public enum Colors {
+	RED, ORANGE, YELLOW, GREEN, BLUE, INDIGO, VIOLET, RAINBOW, PINK
+}
