+++ Laser_Tag/src/mc/cyberplex/LaserTag/kits/enumerations/CoolDownTime.java
@@ -1,0 +1,5 @@
+package mc.cyberplex.LaserTag.kits.enumerations;
+
+public enum CoolDownTime {
+	TEN, NINE, EIGHT, SEVEN, SIX, FIVE, FOUR, THREE, TWO
+}
