+++ Laser_Tag/src/mc/cyberplex/LaserTag/kits/enumerations/GunRanges.java
@@ -1,0 +1,5 @@
+package mc.cyberplex.LaserTag.kits.enumerations;
+
+public enum GunRanges {
+	FIVE, TEN, FIFTEEN, TWENTY, TWENTYFIVE
+}
