+++ Laser_Tag/src/mc/cyberplex/LaserTag/kits/enumerations/GunTypes.java
@@ -1,0 +1,5 @@
+package mc.cyberplex.LaserTag.kits.enumerations;
+
+public enum GunTypes{
+	WOOD, STONE, IRON, GOLD, DIAMOND
+}
