--- Laser_Tag/src/mc/cyberplex/us/arena/Message.java
@@ -1,5 +1,0 @@
-package mc.cyberplex.us.arena;
-
-public enum Message {
-	LOBBY, GAME
-}
