--- Laser_Tag/src/mc/cyberplex/us/kits/enumerations/Colors.java
@@ -1,5 +1,0 @@
-package mc.cyberplex.us.kits.enumerations;
-
-public enum Colors {
-	RED, ORANGE, YELLOW, GREEN, BLUE, INDIGO, VIOLET, RAINBOW, PINK
-}
