--- Laser_Tag/src/mc/cyberplex/us/kits/enumerations/CoolDownTime.java
@@ -1,5 +1,0 @@
-package mc.cyberplex.us.kits.enumerations;
-
-public enum CoolDownTime {
-	TEN, NINE, EIGHT, SEVEN, SIX, FIVE, FOUR, THREE, TWO
-}
