--- Laser_Tag/src/mc/cyberplex/us/kits/enumerations/GunRanges.java
@@ -1,5 +1,0 @@
-package mc.cyberplex.us.kits.enumerations;
-
-public enum GunRanges {
-	FIVE, TEN, FIFTEEN, TWENTY, TWENTYFIVE
-}
