--- Laser_Tag/src/mc/cyberplex/us/kits/enumerations/GunTypes.java
@@ -1,5 +1,0 @@
-package mc.cyberplex.us.kits.enumerations;
-
-public enum GunTypes{
-	WOOD, STONE, IRON, GOLD, DIAMOND
-}
